a lot of new parts and changes (selectors in event, changing event price point, errors views, my account controller and many other)

+++ src/Controller/GiftController.php
@@ -18,20 +18,21 @@ class GiftController extends AbstractController
     #[Route('/', name: 'app_gift_index', methods: ['GET'])]
     public function index(GiftRepository $giftRepository): Response
     {
-        return $this->render('gift/index.html.twig', [
-            'gifts' => $giftRepository->findAll(),
-        ]);
+        return $this->redirectToRoute('index', [], Response::HTTP_SEE_OTHER);
     }
 
     #[Route('/new/{eventId}', name: 'app_gift_new', methods: ['GET', 'POST'])]
     public function new(Request $request, GiftRepository $giftRepository, EventRepository $eventRepository, $eventId, EventUtil $eventUtil): Response
     {
         $event = $eventRepository->find($eventId);
-        if (!$event->isIsPaid() or !$eventUtil->isEventActive($event)){
-            return $this->redirectToRoute('app_event_show', ['id'=> $eventId], Response::HTTP_SEE_OTHER);
+        if (!$event or !$eventId){
+            throw $this->createNotFoundException('Wydarzenie nie istnieje');
         }
-        if ($event->getAmountOfGifts()>=$event->getPricePoint()->getAmountOfGifts()){
-            return $this->redirectToRoute('app_event_show', ['id'=> $eventId], Response::HTTP_SEE_OTHER);
+        if (!$event->isIsPaid() or !$eventUtil->isEventActive($event) or !$eventUtil->isUserACreator($event)){
+            return $this->redirectToRoute('app_event_show', ['selector'=> $event->getSelector()], Response::HTTP_SEE_OTHER);
+        }
+        if ($event->getAmountOfGifts() >= $event->getPricePoint()->getAmountOfGifts()){
+            return $this->redirectToRoute('app_event_show', ['selector'=> $event->getSelector()], Response::HTTP_SEE_OTHER);
         }
 
         $gift = new Gift();
@@ -42,7 +43,7 @@ class GiftController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $giftRepository->save($gift, true);
 
-            return $this->redirectToRoute('app_event_show', ['id'=> $eventId], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_event_show', ['selector'=> $event->getSelector()], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('gift/new.html.twig', [
@@ -55,9 +56,7 @@ class GiftController extends AbstractController
     #[Route('/{id}', name: 'app_gift_show', methods: ['GET'])]
     public function show(Gift $gift): Response
     {
-        return $this->render('gift/show.html.twig', [
-            'gift' => $gift,
-        ]);
+        return $this->redirectToRoute('app_event_show', ['id'=> $gift->getEvent()->getId()], Response::HTTP_SEE_OTHER);
     }
 
     #[Route('/{id}/edit', name: 'app_gift_edit', methods: ['GET', 'POST'])]
@@ -71,7 +70,7 @@ class GiftController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $giftRepository->save($gift, true);
 
-            return $this->redirectToRoute('app_event_show', ['id'=> $event->getId()], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_event_show', ['selector'=> $event->getSelector()], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('gift/edit.html.twig', [
@@ -84,10 +83,17 @@ class GiftController extends AbstractController
     #[Route('/{id}', name: 'app_gift_delete', methods: ['POST'])]
     public function delete(Request $request, Gift $gift, GiftRepository $giftRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$gift->getId(), $request->request->get('_token'))) {
-            $giftRepository->remove($gift, true);
-        }
+        $event = $gift->getEvent();
 
-        return $this->redirectToRoute('app_gift_index', [], Response::HTTP_SEE_OTHER);
+        if ($gift->getDonations()){
+            $this->addFlash('fail', 'Niestety nie możesz usunąć prezentu który posiada już wpłaty');
+        }
+        else {
+            if ($this->isCsrfTokenValid('delete'.$gift->getId(), $request->request->get('_token'))) {
+                $giftRepository->remove($gift, true);
+            }
+        }
+        return $this->redirectToRoute('app_event_show', ['selector'=> $event->getSelector()], Response::HTTP_SEE_OTHER);
+
     }
 }
